+++ libs/zhi-rubick-core/src/browsers/runner.ts
@@ -24,5 +24,5 @@
  */
 
 export default () => {
-  let view;
+  let view
 }